fix callback prototype

+++ api/oc_knx_client.h
@@ -27,7 +27,7 @@ extern "C" {
 
 typedef void (*oc_spake_cb_t)(int error, char *serial_number, char *oscore_id,
                               uint8_t* secret,
-                              uint8_t secret_size);
+                              int secret_size);
 
 /**
  * @brief set the spake response callback
